feat(vhvu-050): invert data flow — self-fetch visuals → props-in + split ViewDataService
Refactor the 3 self-fetching VisualHost visuals (Calendar, DueDateCard,
DueDateCardList) into pure presentational components in @spaarke/visuals plus
thin PCF-side data-fetching containers. Split ViewDataService into pure
FetchXML-string builders vs WebAPI executors.

- @spaarke/visuals: new pure CalendarVisual (events + detailedEvents +
  fetchError props), DueDateCard (cardProps + loading/error), DueDateCardList
  (cards + loading/error + navigatingId). NFR-05: zero webApi/Xrm/FetchXML.
- PCF containers keep the existing export names (CalendarVisual /
  DueDateCardVisual / DueDateCardListVisual), own the fetch + record mapping +
  window.Xrm navigation, and render the pure components. ChartRenderer unchanged.
- New control/services/fetchXmlBuilders.ts holds the pure string helpers
  (injectContextFilter/injectRequiredAttributes/applyMaxItems/substituteParameters);
  ViewDataService re-exports them so DataAggregationService + containers keep
  their import paths.

Directional deviation from POML step 1 (per CLAUDE.md §8.5): pure FetchXML
helpers stay PCF-side, not in @spaarke/visuals, to honor the binding
"@spaarke/visuals is presentational only — no FetchXML" non-negotiable; nothing
in the package consumes them after the refactor.

Behavior-preserving (owner-confirmed): changed WHERE the fetch happens, not
WHAT renders. Verified: @spaarke/visuals tsc green; build:prod green (bundle
762 KiB, leak-free, cleanGuid + footer v1.4.36 intact); service tests 48/48.
Step 9.5 code-review: clean (no Critical/Warning). Component-test harness
blocked by pre-existing @testing-library TS2305 → pure-visual tests deferred to
VHVU-090.
